+++ src/main/java/Nature.java
@@ -1,8 +1,8 @@
-import java.util.Scanner;
-
 public class Nature {
 
     private String message;
+    private final String[] taskList = new String[100];
+    private int taskCount = 0;
 
     public String greeting() {
         return "Hello! I'm Nature.\n"
@@ -14,16 +14,19 @@ public class Nature {
                 + "Hope to see you again soon!";
     }
 
-    public static void main(String[] args) {
-        Nature chatbot = new Nature();
-        System.out.println(chatbot.greeting());
-        Scanner sc = new Scanner(System.in);
-        while(sc.hasNext()) {
-            String s = sc.nextLine();
-            if (s.equals("bye")) break;
-            System.out.println(s);
+    public void addTask(String task) {
+        if (taskCount <= 100) {
+            taskList[taskCount] = task;
+            taskCount++;
         }
-        sc.close();
-        System.out.println(chatbot.farewell());
+        else System.out.println("Can't add more tasks!");
     }
+
+    public void printTaskList() {
+        for (int i = 0; i < taskCount; i++) {
+            int index = i + 1;
+            System.out.println(index + ". " + taskList[i]);
+        }
+    }
+
 }
